Move site-wide construct_whitelister_element_rules Wagtail hook to v1
